Refactor validation code
Add validations for sagas
