+++ language/src/test/scala/com/reactific/riddl/language/ValidationTest.scala
@@ -8,7 +8,7 @@ package com.reactific.riddl.language
 
 import com.reactific.riddl.language.AST.*
 import com.reactific.riddl.language.Messages.*
-import com.reactific.riddl.language.Validation.ValidationState
+import com.reactific.riddl.language.validation.ValidationState
 import com.reactific.riddl.language.ast.At
 import com.reactific.riddl.language.parsing.RiddlParserInput
